Refactoring classes in different files
