+++ Source/Machine.Specifications/ComparisionResult.cs
@@ -1,0 +1,18 @@
+﻿namespace Machine.Specifications
+{
+    public class ComparisionResult
+    {
+        public bool FoundResult { get; set; }
+        public int Result { get; set; }
+
+        public ComparisionResult(int result)
+        {
+            this.FoundResult = true;
+            this.Result = result;
+        }
+
+        protected ComparisionResult()
+        {
+        }
+    }
+}
