+++ Source/Machine.Specifications/DefaultComparer.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace Machine.Specifications
+{
+    class DefaultComparer<T> : IComparer<T>
+    {
+        public int Compare(T x, T y)
+        {
+            // Last case, rely on Object.Equals
+            return object.Equals(x, y) ? 0 : -1;
+        }
+    }
+}
