+++ Source/Machine.Specifications/IComparerStrategy.cs
@@ -1,1 +1,7 @@
-﻿
+﻿namespace Machine.Specifications
+{
+    public interface IComparerStrategy<T>
+    {
+        ComparisionResult Compare(T x, T y);
+    }
+}
