+++ Source/Machine.Specifications/Machine.Specifications.csproj
@@ -71,13 +71,22 @@
     <Compile Include="AssertComparer.cs" />
     <Compile Include="BehaviorsAttribute.cs" />
     <Compile Include="Catch.cs" />
+    <Compile Include="ComparableComparer.cs" />
+    <Compile Include="ComparerFactory.cs" />
+    <Compile Include="ComparisionResult.cs" />
+    <Compile Include="DefaultComparer.cs" />
+    <Compile Include="EnumerableComparer.cs" />
+    <Compile Include="EquatableComparer.cs" />
     <Compile Include="ExtensionMethods.cs" />
     <Compile Include="Factories\BehaviorFactory.cs" />
     <Compile Include="Factories\ITagExtractor.cs" />
+    <Compile Include="GenericTypeComparer.cs" />
     <Compile Include="ICleanupAfterEveryContextInAssembly.cs" />
     <Compile Include="IComparerStrategy.cs" />
     <Compile Include="Model\Behavior.cs" />
     <Compile Include="Model\BehaviorSpecification.cs" />
+    <Compile Include="NoResult.cs" />
+    <Compile Include="ObjectExtension.cs" />
     <Compile Include="Runner\Impl\AssemblyRunner.cs" />
     <Compile Include="Runner\Impl\ContextRunnerFactory.cs" />
     <Compile Include="Runner\Impl\RunListenerBase.cs" />
@@ -110,6 +119,8 @@
     <Compile Include="Runner\SpecificationInfo.cs" />
     <Compile Include="Runner\Impl\DefaultRunner.cs" />
     <Compile Include="TagsAttribute.cs" />
+    <Compile Include="TypeComparer.cs" />
+    <Compile Include="TypeExtension.cs" />
     <Compile Include="Utility\ConsoleRedirection.cs" />
     <Compile Include="Utility\ConsoleStreams.cs" />
     <Compile Include="Utility\Internal\Naming.cs" />
