+++ Source/Machine.Specifications/NoResult.cs
@@ -1,0 +1,10 @@
+﻿namespace Machine.Specifications
+{
+    public class NoResult : ComparisionResult
+    {
+        public NoResult() : base()
+        {
+            this.FoundResult = false;
+        }
+    }
+}
