+++ Source/Machine.Specifications/ObjectExtension.cs
@@ -1,0 +1,10 @@
+﻿namespace Machine.Specifications
+{
+    public static class ObjectExtension
+    {
+        public static bool IsEqualToDefault<T>(this T obj)
+        {
+            return object.Equals(obj, default(T));
+        }
+    }
+}
